+++ prison-mines/src/main/java/tech/mcprison/prison/mines/legacy/MinesFilter.java
@@ -19,15 +19,15 @@
 package tech.mcprison.prison.mines.legacy;
 
 import tech.mcprison.prison.mines.data.Mine;
-import tech.mcprison.prison.mines.managers.LegacyMineManager;
+import tech.mcprison.prison.mines.managers.MineManager;
 
 /**
- * A filter usable with certain methods in {@link LegacyMineManager}
+ * A filter usable with certain methods in {@link MineManager}
  */
 public interface MinesFilter {
 
     /**
-     * Used in {@link LegacyMineManager#select(MinesFilter)}
+     * Used in {@link MineManager#select(MinesFilter)}
      * <p>
      * Check if this {@link MinesFilter} accepts the {@link Mine}. If it does, continue operating.
      *
@@ -37,7 +37,7 @@ public interface MinesFilter {
     boolean accept(Mine c);
 
     /**
-     * Used in {@link LegacyMineManager#forEach(MinesFilter)}
+     * Used in {@link MineManager#forEach(MinesFilter)}
      * <p>
      * Execute methods related to the specified {@link Mine}.
      *
